+++ packages/eas-cli/src/graphql/generated.ts
@@ -79,6 +79,8 @@ export type Account = {
   __typename?: 'Account';
   /** @deprecated Legacy access tokens are deprecated */
   accessTokens: Array<Maybe<AccessToken>>;
+  /** Server account feature gate values for this account, optionally filtering by desired gates. */
+  accountFeatureGates: Scalars['JSONObject']['output'];
   /** Coalesced project activity for all apps belonging to this account. */
   activityTimelineProjectActivities: Array<ActivityTimelineProjectActivity>;
   appCount: Scalars['Int']['output'];
@@ -116,10 +118,6 @@ export type Account = {
   /** Billing information. Only visible to members with the ADMIN or OWNER role. */
   billing?: Maybe<Billing>;
   billingPeriod: BillingPeriod;
-  /** (EAS Build) Builds associated with this account */
-  builds: Array<Build>;
-  /** Whether this account can enable SSO. */
-  canEnableSSO: Scalars['Boolean']['output'];
   createdAt: Scalars['DateTime']['output'];
   displayName?: Maybe<Scalars['String']['output']>;
   /** Environment secrets for an account */
@@ -188,6 +186,15 @@ export type Account = {
   viewerUserPermission: UserPermission;
   /** @deprecated Build packs are no longer supported */
   willAutoRenewBuilds?: Maybe<Scalars['Boolean']['output']>;
+};
+
+
+/**
+ * An account is a container owning projects, credentials, billing and other organization
+ * data and settings. Actors may own and be members of accounts.
+ */
+export type AccountAccountFeatureGatesArgs = {
+  filter?: InputMaybe<Array<Scalars['String']['input']>>;
 };
 
 
@@ -359,18 +366,6 @@ export type AccountAuditLogsPaginatedArgs = {
  */
 export type AccountBillingPeriodArgs = {
   date: Scalars['DateTime']['input'];
-};
-
-
-/**
- * An account is a container owning projects, credentials, billing and other organization
- * data and settings. Actors may own and be members of accounts.
- */
-export type AccountBuildsArgs = {
-  limit: Scalars['Int']['input'];
-  offset: Scalars['Int']['input'];
-  platform?: InputMaybe<AppPlatform>;
-  status?: InputMaybe<BuildStatus>;
 };
 
 
@@ -2419,6 +2414,16 @@ export type AppleDistributionCertificateMutationDeleteAppleDistributionCertifica
   id: Scalars['ID']['input'];
 };
 
+export type AppleDistributionCertificateQuery = {
+  __typename?: 'AppleDistributionCertificateQuery';
+  byId: AppleDistributionCertificate;
+};
+
+
+export type AppleDistributionCertificateQueryByIdArgs = {
+  id: Scalars['ID']['input'];
+};
+
 export type AppleProvisioningProfile = {
   __typename?: 'AppleProvisioningProfile';
   account: Account;
@@ -2472,6 +2477,16 @@ export type AppleProvisioningProfileMutationDeleteAppleProvisioningProfilesArgs 
 
 export type AppleProvisioningProfileMutationUpdateAppleProvisioningProfileArgs = {
   appleProvisioningProfileInput: AppleProvisioningProfileInput;
+  id: Scalars['ID']['input'];
+};
+
+export type AppleProvisioningProfileQuery = {
+  __typename?: 'AppleProvisioningProfileQuery';
+  byId: AppleProvisioningProfile;
+};
+
+
+export type AppleProvisioningProfileQueryByIdArgs = {
   id: Scalars['ID']['input'];
 };
 
@@ -2893,8 +2908,6 @@ export type Build = ActivityTimelineProjectActivity & BuildOrBuildJob & {
   /** Queue position is 1-indexed */
   initialQueuePosition?: Maybe<Scalars['Int']['output']>;
   initiatingActor?: Maybe<Actor>;
-  /** @deprecated User type is deprecated */
-  initiatingUser?: Maybe<User>;
   iosEnterpriseProvisioning?: Maybe<BuildIosEnterpriseProvisioning>;
   isForIosSimulator: Scalars['Boolean']['output'];
   isGitWorkingTreeDirty?: Maybe<Scalars['Boolean']['output']>;
@@ -2963,6 +2976,7 @@ export type BuildAnnotation = {
   regexFlags?: Maybe<Scalars['String']['output']>;
   regexString: Scalars['String']['output'];
   title: Scalars['String']['output'];
+  type: BuildAnnotationType;
 };
 
 export type BuildAnnotationDataInput = {
@@ -2973,6 +2987,7 @@ export type BuildAnnotationDataInput = {
   regexFlags?: InputMaybe<Scalars['String']['input']>;
   regexString: Scalars['String']['input'];
   title: Scalars['String']['input'];
+  type: BuildAnnotationType;
 };
 
 export type BuildAnnotationFiltersInput = {
@@ -3004,6 +3019,12 @@ export type BuildAnnotationMutationUpdateBuildAnnotationArgs = {
   buildAnnotationData: BuildAnnotationDataInput;
   buildAnnotationId: Scalars['ID']['input'];
 };
+
+export enum BuildAnnotationType {
+  Error = 'ERROR',
+  Info = 'INFO',
+  Warning = 'WARNING'
+}
 
 export type BuildAnnotationsQuery = {
   __typename?: 'BuildAnnotationsQuery';
@@ -4208,6 +4229,7 @@ export enum EnvironmentSecretType {
 
 export type EnvironmentVariable = {
   __typename?: 'EnvironmentVariable';
+  /** @deprecated Environment variables are automatically linked to all apps */
   apps: Array<App>;
   createdAt: Scalars['DateTime']['output'];
   environments?: Maybe<Array<EnvironmentVariableEnvironment>>;
@@ -4254,12 +4276,6 @@ export type EnvironmentVariableMutation = {
   deleteBulkEnvironmentVariables: Array<DeleteEnvironmentVariableResult>;
   /** Delete an environment variable */
   deleteEnvironmentVariable: DeleteEnvironmentVariableResult;
-  /** Bulk link shared environment variables */
-  linkBulkSharedEnvironmentVariables: Array<EnvironmentVariable>;
-  /** Link shared environment variable */
-  linkSharedEnvironmentVariable: EnvironmentVariable;
-  /** Unlink shared environment variable */
-  unlinkSharedEnvironmentVariable: EnvironmentVariable;
   /** Bulk update environment variables */
   updateBulkEnvironmentVariables: Array<EnvironmentVariable>;
   /** Update an environment variable */
@@ -4298,25 +4314,6 @@ export type EnvironmentVariableMutationDeleteBulkEnvironmentVariablesArgs = {
 
 export type EnvironmentVariableMutationDeleteEnvironmentVariableArgs = {
   id: Scalars['ID']['input'];
-};
-
-
-export type EnvironmentVariableMutationLinkBulkSharedEnvironmentVariablesArgs = {
-  linkData: Array<LinkSharedEnvironmentVariableInput>;
-};
-
-
-export type EnvironmentVariableMutationLinkSharedEnvironmentVariableArgs = {
-  appId: Scalars['ID']['input'];
-  environment?: InputMaybe<EnvironmentVariableEnvironment>;
-  environmentVariableId: Scalars['ID']['input'];
-};
-
-
-export type EnvironmentVariableMutationUnlinkSharedEnvironmentVariableArgs = {
-  appId: Scalars['ID']['input'];
-  environment?: InputMaybe<EnvironmentVariableEnvironment>;
-  environmentVariableId: Scalars['ID']['input'];
 };
 
 
@@ -5465,12 +5462,6 @@ export type LinkSentryInstallationToExpoAccountInput = {
   sentryOrgSlug: Scalars['String']['input'];
 };
 
-export type LinkSharedEnvironmentVariableInput = {
-  appId: Scalars['ID']['input'];
-  environment?: InputMaybe<EnvironmentVariableEnvironment>;
-  environmentVariableId: Scalars['ID']['input'];
-};
-
 export type LocalBuildArchiveSourceInput = {
   bucketKey: Scalars['String']['input'];
   type: LocalBuildArchiveSourceType;
@@ -6297,6 +6288,10 @@ export type RootQuery = {
   appStoreConnectApiKey: AppStoreConnectApiKeyQuery;
   /** Top-level query object for querying Apple Device registration requests. */
   appleDeviceRegistrationRequest: AppleDeviceRegistrationRequestQuery;
+  /** Top-level query object for querying Apple distribution certificates. */
+  appleDistributionCertificate?: Maybe<AppleDistributionCertificateQuery>;
+  /** Top-level query object for querying Apple provisioning profiles. */
+  appleProvisioningProfile?: Maybe<AppleProvisioningProfileQuery>;
   /** Top-level query object for querying Apple Teams. */
   appleTeam: AppleTeamQuery;
   asset: AssetQuery;
@@ -6833,6 +6828,7 @@ export enum StatuspageServiceName {
   EasBuild = 'EAS_BUILD',
   EasSubmit = 'EAS_SUBMIT',
   EasUpdate = 'EAS_UPDATE',
+  EasWorkflows = 'EAS_WORKFLOWS',
   GithubApiRequests = 'GITHUB_API_REQUESTS',
   GithubWebhooks = 'GITHUB_WEBHOOKS'
 }
@@ -8686,8 +8682,14 @@ export type WorkflowArtifact = {
   id: Scalars['ID']['output'];
   jobRun: JobRun;
   name: Scalars['String']['output'];
+  storageType: WorkflowArtifactStorageType;
   updatedAt: Scalars['DateTime']['output'];
 };
+
+export enum WorkflowArtifactStorageType {
+  Gcs = 'GCS',
+  R2 = 'R2'
+}
 
 export type WorkflowJob = {
   __typename?: 'WorkflowJob';
@@ -9569,24 +9571,6 @@ export type DeleteEnvironmentSecretMutationVariables = Exact<{
 
 
 export type DeleteEnvironmentSecretMutation = { __typename?: 'RootMutation', environmentSecret: { __typename?: 'EnvironmentSecretMutation', deleteEnvironmentSecret: { __typename?: 'DeleteEnvironmentSecretResult', id: string } } };
-
-export type LinkSharedEnvironmentVariableMutationVariables = Exact<{
-  appId: Scalars['ID']['input'];
-  environment?: InputMaybe<EnvironmentVariableEnvironment>;
-  environmentVariableId: Scalars['ID']['input'];
-}>;
-
-
-export type LinkSharedEnvironmentVariableMutation = { __typename?: 'RootMutation', environmentVariable: { __typename?: 'EnvironmentVariableMutation', linkSharedEnvironmentVariable: { __typename?: 'EnvironmentVariable', id: string, name: string, value?: string | null, environments?: Array<EnvironmentVariableEnvironment> | null, createdAt: any, updatedAt: any, scope: EnvironmentVariableScope, visibility?: EnvironmentVariableVisibility | null, type: EnvironmentSecretType } } };
-
-export type UnlinkSharedEnvironmentVariableMutationVariables = Exact<{
-  appId: Scalars['ID']['input'];
-  environment?: InputMaybe<EnvironmentVariableEnvironment>;
-  environmentVariableId: Scalars['ID']['input'];
-}>;
-
-
-export type UnlinkSharedEnvironmentVariableMutation = { __typename?: 'RootMutation', environmentVariable: { __typename?: 'EnvironmentVariableMutation', unlinkSharedEnvironmentVariable: { __typename?: 'EnvironmentVariable', id: string, name: string, value?: string | null, environments?: Array<EnvironmentVariableEnvironment> | null, createdAt: any, updatedAt: any, scope: EnvironmentVariableScope, visibility?: EnvironmentVariableVisibility | null, type: EnvironmentSecretType } } };
 
 export type CreateEnvironmentVariableForAccountMutationVariables = Exact<{
   input: CreateSharedEnvironmentVariableInput;